+++ src/telegram.js
@@ -51,9 +51,12 @@ class TelegramBot extends EventEmitter {
    * @param {Object} [options]
    * @param {Boolean|Object} [options.polling=false] Set true to enable polling or set options.
    *  If a WebHook has been set, it will be deleted automatically.
-   * @param {String|Number} [options.polling.timeout=10] Timeout in seconds for long polling
+   * @param {String|Number} [options.polling.timeout=10] *Deprecated. Use `options.polling.params` instead*.
+   *  Timeout in seconds for long polling.
    * @param {String|Number} [options.polling.interval=300] Interval between requests in miliseconds
    * @param {Boolean} [options.polling.autoStart=true] Start polling immediately
+   * @param {Object} [options.polling.params] Parameters to be used in polling API requests.
+   *  See https://core.telegram.org/bots/api#getupdates for more information.
    * @param {Boolean|Object} [options.webHook=false] Set true to enable WebHook or set options
    * @param {String} [options.webHook.host=0.0.0.0] Host to bind to
    * @param {Number} [options.webHook.port=8443] Port to bind to